refactor: decouple prisma metrics from shared library
